Move relation test to Relation folder
