+++ test/Relation/Stub/StubTableSakura.php
@@ -6,7 +6,7 @@
  * @license    GNU General Public License version 2 or later;
  */
 
-namespace Windwalker\Test\Table\Stub;
+namespace Windwalker\Test\Relation\Stub;
 
 use Windwalker\Table\Table;
